backend/requirements.txt
fastapi
uvicorn
python-multipart
python-redis
redis
pillow
qrcode
sqlalchemy
aiofiles
pydantic
pydantic-settings
python-dotenv
celery
requests
uvloop
httpx
pandas
openpyxl
passlib[bcrypt]
python-jose[cryptography]
python-multipart
